+++ test/training/medium/telephoneNumbers/CGTest.js
@@ -1,0 +1,75 @@
+import { assert } from 'chai';
+import sinon from 'sinon';
+import File from '../../../File.js';
+import { execute } from '../../../../lib/training/medium/telephoneNumbers/telephoneNumbers.js';
+
+/**
+ * Tests for the "Telephone numbers" puzzle.
+ */
+suite('Telephone numbers', function() {
+    const sandbox = sinon.createSandbox();
+
+    setup(function () {
+        sandbox.stub(console, "log");
+    });
+
+    teardown(function () {
+        sandbox.restore();
+    });
+
+
+    test('A telephone number', function() {
+        let inputFile = new File('./test/training/medium/telephoneNumbers/input/01 - a telephone number.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            10
+        );
+    });
+
+    test('Numbers with a different base', function() {
+        let inputFile = new File('./test/training/medium/telephoneNumbers/input/02 - numbers with a different base.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            20
+        );
+    });
+
+    test('Number included in another', function() {
+        let inputFile = new File('./test/training/medium/telephoneNumbers/input/03 - number included in another.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            10
+        );
+    });
+
+    test('Numbers with a common part', function() {
+        let inputFile = new File('./test/training/medium/telephoneNumbers/input/04 - numbers with a common part.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            28
+        );
+    });
+
+    test('Long list of numbers', function() {
+        let inputFile = new File('./test/training/medium/telephoneNumbers/input/05 - long list of numbers.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            45512
+        );
+    });
+});