+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Network cabling".
 - Tests for "Scrabble".
 - Tests for "Stock exchange losses".
+- Tests for "Telephone numbers".
 
 ## [1.2.0] - 2022-03-02
 ### Added